Add ark plugins - fixes
