+++ tools/ark-plugins/Gopkg.toml
@@ -69,4 +69,4 @@ required = [
 
 [[constraint]]
   name = "github.com/kubernetes-incubator/service-catalog"
-  version = "=v0.1.28"
+  version = "=v0.1.34"
